+++ System/Monopati/Posix.hs
@@ -1,4 +1,6 @@
 module System.Monopati.Posix (
+	-- * Types and instances
+	module System.Monopati.Posix.Core,
 	-- * CRUD operations for filesystem
 	module System.Monopati.Posix.Calls,
 	-- * Pure combinators
@@ -7,3 +9,4 @@ module System.Monopati.Posix (
 
 import System.Monopati.Posix.Calls
 import System.Monopati.Posix.Combinators
+import System.Monopati.Posix.Core