+++ System/Monopati/Posix/Calls.hs
@@ -14,7 +14,8 @@ import "directory" System.Directory (createDirectoryIfMissing, getCurrentDirecto
 import "free" Control.Comonad.Cofree (Cofree ((:<)))
 import "split" Data.List.Split (splitOn)
 
-import System.Monopati.Posix.Combinators (Points (Directory), To, Path, Origin (Root), Outline (Outline), Absolute, Homeward, Relative)
+import System.Monopati.Posix.Combinators (Absolute, Homeward, Relative)
+import System.Monopati.Posix.Core (Points (Directory), To, Path, Origin (Root), Outline (Outline))
 
 -- | Return Nothing, if current working directory is root (cwd)
 current :: IO (Maybe (Absolute Path To Directory))